requirements.txt
cffi==1.16.0
numpy==1.26.3
pillow==10.2.0
pycparser==2.21
